# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Paperboat Server API",
    "version": "0.11.0",
    "description": "Production control-plane HTTP API for Paperboat dashboard and CLI clients."
  },
  "servers": [
    {
      "url": "{base_url}",
      "variables": {
        "base_url": {
          "default": "http://127.0.0.1:8080"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "paperboat_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "bearerAccess": {
        "type": "http",
        "scheme": "bearer",
        "description": "Short-lived scoped Paperboat access token."
      },
      "bearerRefresh": {
        "type": "http",
        "scheme": "bearer",
        "description": "Rotating refresh token accepted only by the refresh endpoint."
      },
      "bearerEdgeControl": {
        "type": "http",
        "scheme": "bearer",
        "description": "Private edge-control service credential. Never exposed to browser or CLI clients."
      },
      "bearerHelperIdentity": {
        "type": "http",
        "scheme": "bearer",
        "description": "Short-lived helper identity credential bound to the helper Ed25519 key."
      },
      "bearerPreviewRegistration": {
        "type": "http",
        "scheme": "bearer",
        "description": "Five-minute preview_registration credential bound to one helper and environment."
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {}
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        }
      },
      "APIError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "examples": [
              "machine_not_ready"
            ]
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "EnvironmentConnectionDescriptor": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "issuer",
          "connectable",
          "expires_at",
          "environment",
          "status",
          "reason",
          "retry_after_seconds"
        ],
        "properties": {
          "schema": {
            "const": "paperboat.environment-connection/v1"
          },
          "issuer": {
            "type": "string",
            "format": "uri",
            "pattern": "^https://"
          },
          "connectable": {
            "type": "boolean"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "environment": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "id",
              "kind",
              "resource_id",
              "display_name",
              "state"
            ],
            "properties": {
              "id": {
                "type": "string",
                "minLength": 1
              },
              "kind": {
                "type": "string",
                "enum": [
                  "hosted",
                  "byod"
                ]
              },
              "resource_id": {
                "type": "string",
                "minLength": 1
              },
              "display_name": {
                "type": "string",
                "minLength": 1
              },
              "state": {
                "type": "string",
                "enum": [
                  "starting",
                  "ready",
                  "offline",
                  "revoked",
                  "deleted"
                ]
              },
              "root": {
                "type": "string",
                "minLength": 1
              }
            }
          },
          "helper": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "id",
              "generation"
            ],
            "properties": {
              "id": {
                "type": "string",
                "minLength": 1
              },
              "generation": {
                "type": "integer",
                "minimum": 1
              }
            }
          },
          "capabilities": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": [
                "terminal",
                "herdr",
                "upload",
                "preview",
                "activity"
              ]
            }
          },
          "terminal": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "endpoint",
              "auth",
              "session_id",
              "thread_id",
              "terminal_id",
              "cwd"
            ],
            "properties": {
              "endpoint": {
                "type": "string",
                "format": "uri",
                "pattern": "^wss://"
              },
              "http_endpoint": {
                "type": "string",
                "format": "uri",
                "pattern": "^https://"
              },
              "auth": {
                "$ref": "#/components/schemas/TerminalAuth"
              },
              "session_id": {
                "type": "string",
                "minLength": 1
              },
              "thread_id": {
                "type": "string",
                "minLength": 1
              },
              "terminal_id": {
                "type": "string",
                "minLength": 1
              },
              "cwd": {
                "type": "string",
                "minLength": 1
              }
            }
          },
          "upload": {
            "$ref": "#/components/schemas/StagedImageUpload"
          },
          "status": {
            "type": "string",
            "minLength": 1
          },
          "reason": {
            "type": "string",
            "minLength": 1
          },
          "retry_after_seconds": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "TerminalAuth": {
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "method",
              "ticket",
              "expires_at",
              "scopes"
            ],
            "properties": {
              "method": {
                "const": "websocket_ticket"
              },
              "ticket": {
                "type": "string",
                "minLength": 1
              },
              "expires_at": {
                "type": "string",
                "format": "date-time"
              },
              "scopes": {
                "type": "array",
                "minItems": 1,
                "maxItems": 1,
                "prefixItems": [
                  {
                    "const": "terminal:operate"
                  }
                ],
                "items": false
              }
            }
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "method",
              "token",
              "expires_at",
              "scopes"
            ],
            "properties": {
              "method": {
                "const": "bearer"
              },
              "token": {
                "type": "string",
                "minLength": 1
              },
              "expires_at": {
                "type": "string",
                "format": "date-time"
              },
              "scopes": {
                "type": "array",
                "minItems": 1,
                "maxItems": 1,
                "prefixItems": [
                  {
                    "const": "terminal:operate"
                  }
                ],
                "items": false
              }
            }
          }
        ]
      },
      "StagedImageUpload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "endpoint",
          "auth",
          "max_bytes",
          "allowed_mime_types",
          "retention_seconds"
        ],
        "properties": {
          "endpoint": {
            "type": "string",
            "format": "uri",
            "pattern": "^https://"
          },
          "auth": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "method",
              "token",
              "expires_at",
              "scopes"
            ],
            "properties": {
              "method": {
                "const": "bearer"
              },
              "token": {
                "type": "string",
                "minLength": 1
              },
              "expires_at": {
                "type": "string",
                "format": "date-time"
              },
              "scopes": {
                "type": "array",
                "minItems": 1,
                "maxItems": 1,
                "prefixItems": [
                  {
                    "const": "file:stage"
                  }
                ],
                "items": false
              }
            }
          },
          "max_bytes": {
            "type": "integer",
            "minimum": 1
          },
          "allowed_mime_types": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "pattern": "^image/"
            }
          },
          "retention_seconds": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
      "ConnectionReadiness": {
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "project_id",
              "project_state",
              "connectable",
              "status",
              "reason",
              "retry_after_seconds"
            ],
            "properties": {
              "project_id": {
                "type": "string"
              },
              "project_state": {
                "type": "string"
              },
              "connectable": {
                "const": true
              },
              "status": {
                "const": "ready"
              },
              "reason": {
                "const": "ready"
              },
              "retry_after_seconds": {
                "const": 0
              }
            }
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "project_id",
              "project_state",
              "connectable",
              "status",
              "reason",
              "retry_after_seconds"
            ],
            "properties": {
              "project_id": {
                "type": "string"
              },
              "project_state": {
                "type": "string"
              },
              "connectable": {
                "const": false
              },
              "status": {
                "const": "machine_starting"
              },
              "reason": {
                "type": "string",
                "enum": [
                  "machine_start_queued",
                  "machine_not_running"
                ]
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1
              }
            }
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "project_id",
              "project_state",
              "connectable",
              "status",
              "reason",
              "retry_after_seconds"
            ],
            "properties": {
              "project_id": {
                "type": "string"
              },
              "project_state": {
                "type": "string"
              },
              "connectable": {
                "const": false
              },
              "status": {
                "const": "tunnel_connecting"
              },
              "reason": {
                "const": "tunnel_offline"
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1
              }
            }
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "project_id",
              "project_state",
              "connectable",
              "status",
              "reason",
              "retry_after_seconds"
            ],
            "properties": {
              "project_id": {
                "type": "string"
              },
              "project_state": {
                "type": "string"
              },
              "connectable": {
                "const": false
              },
              "status": {
                "const": "helper_starting"
              },
              "reason": {
                "type": "string",
                "enum": [
                  "helper_unhealthy",
                  "terminal_session_operation_pending"
                ]
              },
              "retry_after_seconds": {
                "type": "integer",
                "minimum": 1
              }
            }
          }
        ]
      },
...
    },
    "responses": {
      "Error": {
        "description": "Structured API error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "parameters": {
      "ProjectID": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    }
  },
  "security": [
    {
      "cookieSession": []
    }
  ],
  "paths": {
...
    "/v1/auth/csrf": {
      "get": {
        "summary": "Read CSRF token for browser writes",
        "responses": {
          "200": {
            "description": "CSRF token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "get_v1_auth_csrf"
      }
    },
...
    "/v1/projects": {
      "get": {
        "summary": "List projects",
        "description": "Cookie authentication requires no bearer scope. Bearer authentication requires projects:read.",
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerAccess": []
          }
        ],
        "x-required-bearer-scopes": [
          "projects:read"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created_at",
                "-created_at",
                "updated_at",
                "-updated_at",
                "name",
                "-name",
                "state",
                "-state"
              ],
              "default": "-created_at"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "get_v1_projects"
      },
      "post": {
        "summary": "Create project",
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerAccess": []
          }
        ],
        "x-required-bearer-scopes": [
          "projects:connect"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "Project created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "post_v1_projects"
      }
    },
...
    "/v1/projects/{project_id}/connection-descriptor": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectID"
        }
      ],
      "post": {
        "summary": "Create CLI access descriptor",
        "description": "Bearer authentication requires projects:connect. The optional body selects a terminal_session_id; omission selects default.",
        "security": [
          {
            "bearerAccess": []
          }
        ],
        "x-required-bearer-scopes": [
          "projects:connect"
        ],
        "responses": {
          "200": {
            "description": "Fresh CLI descriptor.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EnvironmentConnectionDescriptor"
                    }
                  }
                }
              }
            }
          },
          "202": {
            "description": "Project is starting; no credentials are returned.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ConnectionReadiness"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "post_v1_projects_project_id_connection_descriptor"
      }
    },
...
  }
}